requirements.txt
ruff==0.9.1
Scrapy==2.12.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.37
python-dotenv==1.0.1
pydantic-settings==2.7.1